refactor(runner): split BaseRunner (shared) from BaseCudaGraphRunner
Introduce a minimal BaseRunner(ABC) parent that the
cuda-graph runners AND the upcoming EagerRunner share, and reparent
BaseCudaGraphRunner(BaseRunner). This is the class-hierarchy keystone: eager's
prepare() is too different from the capture loop to share via a flag, so eager
gets its own sibling under BaseRunner rather than being a flag on the graph
runner.

BaseRunner (new, runner/base_runner.py) holds only the shared surface:
  - __init__ (model_runner, device, tp/dp/pp sizes, attn_tp, tbo_plugin)
  - warmup() (run-once; delegates to ModelRunner.kernel_warmup, threading the
    runner's _autotune_buffers)
  - _autotune_buffers() hook (default None, None)
  - abstract can_run_graph / load_batch / execute

BaseCudaGraphRunner keeps the capture/shape machinery (_pad_to_bucket + abstract
reserve_batch / prepare / _prepare_one + the buffers/backend declarations). No
behavior change — pure hierarchy extraction; the bodies moved verbatim.

GPU-validated (H200): Qwen3-0.6B graph+eager coherent; DeepSeek-V2-Lite +
flashinfer_cutlass autotune graph+eager coherent; dense flashinfer + fa3
attention unittests 18 passed / 71 subtests, no regressions.
